+++ R/data_refresh.R
@@ -15,50 +15,48 @@ covid19sf_refresh <- function(force = FALSE){
   cat(paste0("\033[4;", 32, "m","Checking for updates for the covid19sf package","\033[0m","\n"))
 
   # covid19sf_age ----
-  cat("covid19sf_age dataset...")
-
-  age_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_age.csv", stringsAsFactors = FALSE) %>%
-    dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
-
-
-  covid19sf_age_old <- covid19sf::covid19sf_age
-
-  if(max(covid19sf_age_old$specimen_collection_date) < max(age_csv$specimen_collection_date) ||
-     nrow(covid19sf_age_old) < nrow(age_csv)){
-    cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
-    updates <- TRUE
-  } else{
-    cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
-  }
+  # cat("covid19sf_age dataset...")
+  #
+  # age_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_age.csv", stringsAsFactors = FALSE) %>%
+  #   dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
+  #                                                    tz = "America/Los_Angeles"),
+  #                 last_updated = lubridate::ymd_hms(last_updated,
+  #                                                   tz = "America/Los_Angeles"))
+  #
+  #
+  # covid19sf_age_old <- covid19sf::covid19sf_age
+  #
+  # if(max(covid19sf_age_old$specimen_collection_date) < max(age_csv$specimen_collection_date) ||
+  #    nrow(covid19sf_age_old) < nrow(age_csv)){
+  #   cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
+  #   updates <- TRUE
+  # } else{
+  #   cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
+  # }
 
   # covid19sf_summary ----
-  cat("covid19sf_summary dataset...")
-  summary_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_summary.csv", stringsAsFactors = FALSE) %>%
-    dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
-
-  covid19sf_summary_old <- covid19sf::covid19sf_summary
-
-  if(max(covid19sf_summary_old$specimen_collection_date) < max(summary_csv$specimen_collection_date) || nrow(covid19sf_summary_old) < nrow(summary_csv)){
-    cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
-    updates <- TRUE
-  } else{
-    cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
-  }
+  # cat("covid19sf_summary dataset...")
+  # summary_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_summary.csv", stringsAsFactors = FALSE) %>%
+  #   dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
+  #                                                    tz = "America/Los_Angeles"),
+  #                 last_updated = lubridate::ymd_hms(last_updated,
+  #                                                   tz = "America/Los_Angeles"))
+  #
+  # covid19sf_summary_old <- covid19sf::covid19sf_summary
+  #
+  # if(max(covid19sf_summary_old$specimen_collection_date) < max(summary_csv$specimen_collection_date) || nrow(covid19sf_summary_old) < nrow(summary_csv)){
+  #   cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
+  #   updates <- TRUE
+  # } else{
+  #   cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
+  # }
 
 
   # covid19sf_tests ----
   cat("covid19sf_tests dataset...")
   tests_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_tests.csv", stringsAsFactors = FALSE) %>%
     dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
+                                                     tz = "America/Los_Angeles"))
 
   covid19sf_tests_old <- covid19sf::covid19sf_tests
 
@@ -98,26 +96,26 @@ covid19sf_refresh <- function(force = FALSE){
 
 
   # covid19sf_demo ----
-  cat("covid19sf_demo...")
-
-  covid19sf_demo_old <- covid19sf::covid19sf_demo
-
-  demo_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_demo.csv", stringsAsFactors = FALSE) %>%
-    dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
-
-
-
-
-  if(max(covid19sf_demo_old$specimen_collection_date) < max(demo_csv$specimen_collection_date) ||
-     nrow(covid19sf_demo_old) < nrow(demo_csv)){
-    cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
-    updates <- TRUE
-  } else{
-    cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
-  }
+  # cat("covid19sf_demo...")
+  #
+  # covid19sf_demo_old <- covid19sf::covid19sf_demo
+  #
+  # demo_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_demo.csv", stringsAsFactors = FALSE) %>%
+  #   dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
+  #                                                    tz = "America/Los_Angeles"),
+  #                 last_updated = lubridate::ymd_hms(last_updated,
+  #                                                   tz = "America/Los_Angeles"))
+  #
+  #
+  #
+  #
+  # if(max(covid19sf_demo_old$specimen_collection_date) < max(demo_csv$specimen_collection_date) ||
+  #    nrow(covid19sf_demo_old) < nrow(demo_csv)){
+  #   cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
+  #   updates <- TRUE
+  # } else{
+  #   cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
+  # }
 
   # covid19sf_housing ----
   cat("covid19sf_housing...")
@@ -177,43 +175,43 @@ covid19sf_refresh <- function(force = FALSE){
   }
 
   # covid19sf_gender ----
-  cat("covid19sf_gender...")
-
-  covid19sf_gender_old <- covid19sf::covid19sf_gender
-
-  gender_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_gender.csv", stringsAsFactors = FALSE) %>%
-    dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
-
-  if(max(covid19sf_gender_old$specimen_collection_date) < max(gender_csv$specimen_collection_date) ||
-     nrow(covid19sf_gender_old) > nrow(gender_csv)){
-    cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
-    updates <- TRUE
-  } else{
-    cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
-  }
+  # cat("covid19sf_gender...")
+  #
+  # covid19sf_gender_old <- covid19sf::covid19sf_gender
+  #
+  # gender_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_gender.csv", stringsAsFactors = FALSE) %>%
+  #   dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
+  #                                                    tz = "America/Los_Angeles"),
+  #                 last_updated = lubridate::ymd_hms(last_updated,
+  #                                                   tz = "America/Los_Angeles"))
+  #
+  # if(max(covid19sf_gender_old$specimen_collection_date) < max(gender_csv$specimen_collection_date) ||
+  #    nrow(covid19sf_gender_old) > nrow(gender_csv)){
+  #   cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
+  #   updates <- TRUE
+  # } else{
+  #   cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
+  # }
 
 
   # covid19sf_homeless ----
-  cat("covid19sf_homeless...")
-
-  covid19sf_homeless_old <- covid19sf::covid19sf_homeless
-
-  homeless_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_homeless.csv", stringsAsFactors = FALSE) %>%
-    dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
-                                                     tz = "America/Los_Angeles"),
-                  last_updated = lubridate::ymd_hms(last_updated,
-                                                    tz = "America/Los_Angeles"))
-
-  if(max(covid19sf_homeless_old$specimen_collection_date) < max(homeless_csv$specimen_collection_date) ||
-     nrow(covid19sf_homeless_old) < nrow(homeless_csv)){
-    cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
-    updates <- TRUE
-  } else{
-    cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
-  }
+  # cat("covid19sf_homeless...")
+  #
+  # covid19sf_homeless_old <- covid19sf::covid19sf_homeless
+  #
+  # homeless_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_homeless.csv", stringsAsFactors = FALSE) %>%
+  #   dplyr::mutate(specimen_collection_date = as.Date(specimen_collection_date,
+  #                                                    tz = "America/Los_Angeles"),
+  #                 last_updated = lubridate::ymd_hms(last_updated,
+  #                                                   tz = "America/Los_Angeles"))
+  #
+  # if(max(covid19sf_homeless_old$specimen_collection_date) < max(homeless_csv$specimen_collection_date) ||
+  #    nrow(covid19sf_homeless_old) < nrow(homeless_csv)){
+  #   cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
+  #   updates <- TRUE
+  # } else{
+  #   cat(paste0("\033[0;", 41, "m","no updates are available","\033[0m","\n"))
+  # }
 
 
   # Vaccine demographic summary ----
@@ -241,11 +239,10 @@ covid19sf_refresh <- function(force = FALSE){
   covid19sf_vaccine_demo_ts_old <- covid19sf::covid19sf_vaccine_demo_ts
 
   covid19sf_vaccine_demo_ts_csv <- utils::read.csv("https://raw.githubusercontent.com/RamiKrispin/covid19sf/master/csv/covid19sf_vaccine_demo_ts.csv", stringsAsFactors = FALSE)
-  covid19sf_vaccine_demo_ts_csv$data_as_of <- lubridate::ymd_hms(covid19sf_vaccine_demo_ts_csv$data_as_of)
-  covid19sf_vaccine_demo_ts_csv$data_loaded_at <- lubridate::ymd_hms(covid19sf_vaccine_demo_ts_csv$data_loaded_at)
+  covid19sf_vaccine_demo_ts_csv$data_as_of <- NULL #lubridate::ymd_hms(covid19sf_vaccine_demo_ts_csv$data_as_of)
+  covid19sf_vaccine_demo_ts_csv$data_loaded_at <- NULL #lubridate::ymd_hms(covid19sf_vaccine_demo_ts_csv$data_loaded_at)
 
-  if(round(as.numeric(max(covid19sf_vaccine_demo_ts_csv$data_loaded_at, na.rm = TRUE))) >
-     round(as.numeric(max(covid19sf_vaccine_demo_ts_old$data_loaded_at, na.rm = TRUE)))){
+  if(nrow(covid19sf_vaccine_demo_ts_old) < nrow(covid19sf_vaccine_demo_ts_csv)){
     cat(paste0("\033[0;", 42, "m","updates available","\033[0m","\n"))
     updates <- TRUE
   } else{